Fix typo in index
This commit fixes a typo in the index file of the 30 Days of React project. The typo was in the line "Creating an objecting with values", which should be "Creating an object with values". This correction improves the readability and clarity of the instructions for the learners. No other changes were made to the file or the project.